requirements.txt
# Core dependencies
PyYAML>=6.0
requests>=2.28.0

# SSH/deployment (optional, for remote deployment features)
# paramiko>=3.0.0
# fabric>=3.0.0
